+++ geom/geojson_feature_collection.go
@@ -29,7 +29,7 @@ type GeoJSONFeature struct {
 }
 
 // UnmarshalJSON implements the encoding/json Unmarshaler interface by
-// unmarshalling a GeoJSON Feature Collection object.
+// unmarshalling a GeoJSON Feature.
 func (f *GeoJSONFeature) UnmarshalJSON(p []byte) error {
 	var topLevel map[string]json.RawMessage
 	if err := json.Unmarshal(p, &topLevel); err != nil {
@@ -123,6 +123,11 @@ func (f GeoJSONFeature) MarshalJSON() ([]byte, error) {
 	if len(f.ForeignMembers) == 0 {
 		return buf, nil
 	}
+	for _, forbidden := range []string{"type", "geometry", "id", "properties"} {
+		if _, ok := f.ForeignMembers[forbidden]; ok {
+			return nil, forbiddenForeignMemberError{forbidden}
+		}
+	}
 	fms, err := json.Marshal(f.ForeignMembers)
 	if err != nil {
 		return nil, err
@@ -136,37 +141,90 @@ func (f GeoJSONFeature) MarshalJSON() ([]byte, error) {
 // GeoJSONFeatureCollection is a collection of GeoJSONFeatures.
 // GeoJSONFeatureCollection values have a one to one correspondence with
 // GeoJSON FeatureCollections.
-type GeoJSONFeatureCollection []GeoJSONFeature
+type GeoJSONFeatureCollection struct {
+	Features       []GeoJSONFeature
+	ForeignMembers map[string]interface{}
+}
 
 // UnmarshalJSON implements the encoding/json Unmarshaler interface by
 // unmarshalling a GeoJSON FeatureCollection object.
 func (c *GeoJSONFeatureCollection) UnmarshalJSON(p []byte) error {
-	var topLevel struct {
-		Type     string           `json:"type"`
-		Features []GeoJSONFeature `json:"features"`
-	}
+	var topLevel map[string]json.RawMessage
 	if err := json.Unmarshal(p, &topLevel); err != nil {
-		return err
+		return fmt.Errorf("not a valid JSON object: %w", err)
 	}
-	if topLevel.Type == "" {
-		return errors.New("feature collection type field missing or empty")
+
+	typeJSON, ok := topLevel["type"]
+	if !ok {
+		return errors.New("feature collection type field missing")
 	}
-	if topLevel.Type != "FeatureCollection" {
-		return fmt.Errorf("type field not set to FeatureCollection: '%s'", topLevel.Type)
+	var typeStr string
+	if err := json.Unmarshal(typeJSON, &typeStr); err != nil {
+		return fmt.Errorf("unmarshalling type field: %w", err)
 	}
-	*c = topLevel.Features
+	if typeStr != "FeatureCollection" {
+		return fmt.Errorf("type field not set to FeatureCollection: '%s'", typeStr)
+	}
+
+	featuresJSON, ok := topLevel["features"]
+	if !ok {
+		return errors.New("feature collection features field missing")
+	}
+	var features []GeoJSONFeature
+	if err := json.Unmarshal(featuresJSON, &features); err != nil {
+		return fmt.Errorf("unmarshalling features field: %w", err)
+	}
+
+	foreignMembers := make(map[string]interface{})
+	for k, vJSON := range topLevel {
+		switch k {
+		case "type", "features":
+			continue
+		default:
+			var v interface{}
+			if err := json.Unmarshal(vJSON, &v); err != nil {
+				return fmt.Errorf("unmarshalling foreign member '%s': %w", k, err)
+			}
+			foreignMembers[k] = v
+		}
+	}
+
+	*c = GeoJSONFeatureCollection{
+		Features:       features,
+		ForeignMembers: foreignMembers,
+	}
 	return nil
 }
 
 // MarshalJSON implements the encoding/json Marshaler interface by marshalling
 // into a GeoJSON FeatureCollection object.
 func (c GeoJSONFeatureCollection) MarshalJSON() ([]byte, error) {
-	var col []GeoJSONFeature = c
-	if col == nil {
-		col = []GeoJSONFeature{}
+	if c.Features == nil {
+		c.Features = []GeoJSONFeature{}
 	}
-	return json.Marshal(struct {
+
+	buf, err := json.Marshal(struct {
 		Type     string           `json:"type"`
 		Features []GeoJSONFeature `json:"features"`
-	}{"FeatureCollection", col})
+	}{"FeatureCollection", c.Features})
+	if err != nil {
+		return nil, err
+	}
+
+	if len(c.ForeignMembers) == 0 {
+		return buf, nil
+	}
+	for _, forbidden := range []string{"type", "features"} {
+		if _, ok := c.ForeignMembers[forbidden]; ok {
+			return nil, forbiddenForeignMemberError{forbidden}
+		}
+	}
+	fms, err := json.Marshal(c.ForeignMembers)
+	if err != nil {
+		return nil, err
+	}
+	buf = buf[:len(buf)-1] // remove trailing '}'
+	buf = append(buf, ',')
+	buf = append(buf, fms[1:]...) // skip leading '{' (must be a JSON object due to construction)
+	return buf, nil
 }